Simple throttler implementation
Currently this is more or less a stub throttler implementation just to
demonstrate the architecture of the throttler, which is a separate go
routine that collects mysql status variables to determine the load, and
then set a throttle timeout if the load exceeds the predefined maximum.

Additionally, it is possible to manually set a throttler timeout on the
web UI.

On each iteration of the DataIterator and the BinlogStreamer, the
throttler is called. This blocks if the throttler is engaged,
effectively killing any queries going to the source and target shard.

It is possible that a manual "unthrottle" is issued, in which case the
throttler will unblock within 500ms.