Ghostferry TLS
This patch enables TLS for Ghostferry's MySQL transport. To do this, we
must switch to the "official" golang mysql driver as opposed to using
the driver provided by the siddontang/go-mysql package (which we use for
things like replication). The reason is that siddontang's driver does
not have SSL support and adding it would be a larger job than to just
switch.

This also removes a hack where we did the CREATE TABLE statements on the
target database via a separate database connection.