Stop using a separate client for schema cache
I was able to merge a PR[1] that allowed the upstream library to get the
table schema using an instance of *sql.DB. This allows us to use the
same connection to fetch the table schema.

Getting rid of this hack is essential for TLS support, as creating
another connection with TLS would be annoying.

[1]: https://github.com/siddontang/go-mysql/pull/151