Fixed case where the max row is deleted
If the max row is deleted and nothing inserted after it, previously,
the data iterator will never finish as it will never reach the max
primary key.

If the select query for the data iterator (which does WHERE id > ...
LIMIT ...) returns no data, it knows that the data set has been
completely copied, as there max primary key has changed.