Build ghostferry-copydb debian package
Since ghostferry consists of a binary and a directory of static assets that
it needs, building a debian package is a simple way to bundle the static
assets with ghostferry.

The alternative is worse, as either we have to pass the directory to the
web UI assets as a config option to ghostferry, be only allowed to
run from one particular directory, or find a way to bundle the static
file into the binary.

As far as I can tell, bundling the static file into the binary is all
slightly non-trivial and adds extra dependices. The top library for this
(go-bindata) is also unmaintained.