Minor fixes.
The throttling code was in the wrong spot. It won't throttle retries in
the previous spot.

The verifier could be started multiple times, which is racy and
redundant. This has been "fixed" as it is still possible to start two
verification attempts, just less likely. However it doesn't matter that
much so this is deemed acceptable.