Measure BinlogStreamer lag
The BinlogStreamer lag is determined as the current time - the last
processed event time. This is relatively easy to obtain as each binlog
event has a timestamp associated to it.

There are two challenges:

1. Some events do not have the timestamp in its header. Notably,
   RotateEvent does not have a timestamp associated. However, it's
   unclear which/if other event also does not have a timestamp as the MySQL
   documentation is unclear about this.
2. If we do not receive a BinlogEvent for an extended period of time, we
   the lag will grow.

For (1), only I've excluded RotateEvent and FormatDescriptionEvent from
being able to set the lastProcessedEventTime.

For (2), every time we failed to get an event (getting a BinlogEvent
will timeout after 500ms), we update the lastProcessedEventTime.

This allows us to somewhat accurately measure the BinlogStreamer's lag.

However, there is a potential issues:

If we fail to get an event but there actually is an event in the past
that is stuck in the queue or something, we could update the
lastProcessedEventTime to a time that's backwards when the stuck
event eventually comes through. There could also be cases where the date
of the ghostferry server is slightly out of sync with the mysql.

I'm not sure if this is an issue, so I'm not dealing with it apart from
emitting it to the logs.