ETA for Ghostferry
To estimate the time required for a Ghostferry run, we simply need to
know how long it will take to iterate through all the tables, as the
remaining work will likely be negligible in comparison.

To perform this estimation, the formula
`ETA = (TotalNumberOfPKs - TotalPKsProcessed)/PKsProcessedPerSecond`
is utilized.

We know:

```
TotalNumberOfPKs = \sum_for all tables (TargetPKs - Minimum PK)
TotalPKsProcessed = \sum_for all tables (LastSuccessfulPK)
```

To calculate `PKsProcessedPerSecond`, we log the total primary keys
processed every time we update the LastSuccessfulPK. This log keeps the
most recent 20 updates and the timestamp of each time we update. From
this, we area able to compute a moving average of the speed.

Together, we have the ETA estimation.