Allow blacklist a table from being copied
Previously, ghostferry operated as a whitelist. This patch enables us to
use either a blacklist or a whitelist.

By setting false for a particular database/table in the
ApplicableDatabase/ApplicableTable entry in the config, you can
explicitly exclude a particular database or a table.

However, to use this feature, you must include `"*": true` in the
config, as otherwise you still need to whitelist individual
tables/databases.