More generic Makefile
Added the date in the deb version string. This is important to
distinguish between the latest release and an older one as commit shas
do not guarentee order.

Also made the Makefile more generic so the building between copydb and
reloc is consistent.

This does mean the package name and binary name of reloc is altered to
be ghostferry-reloc instead of just reloc.