Verifier refactor
Refactored Verifier to have a better interface. Also implemented
DatabaseRewrites and TableRewrites for ChecksumTableVerifier and make it
fail upon the first mismatched table as opposed to everything.

The idea is to eventually move the Verifier out of Ferry all together,
along with the ControlServer. This is merely the first step towards
that.