Buffers for the reverify queues
Without a buffer they may deadlock if many jobs fail and cause the
processing goroutines to exit, and the main thread is stuck sending
things down the channel.

However, the buffer size of the channel must be equal to the size of the
things it is trying to enqueue. This is not trivial to determine for the
reverifyStore. Right now the size is 2000 and this is a short term fix.
Long term there's still a chance of deadlock here.