Added worker pool to address deadlocking issues
Previously, if all workers in the iterative verifier errors before all
the batches/tables gets queued via the channel, the iterative verifier
may deadlock as it tries to send the batches/tables down a channel that
nothing is listening to.

This commit addresses this by introducing a generic WorkerPool struct
that tries to send but fails after 500ms and then checking to see if any
goroutine it spawned has errored. If it has errored, the system will
bail. If not, it'll keep trying to queuing work.