WorkerPool sends error back via a channel
A select for either send work or receive the error is used to exit
early. This simplifies the error handling.

Also instead of passing in an array, passes just an integer which
represents the index in the work array. Each process call gets that
integer instead of an interface to the actual data. Closure is expected
to be used to fetch the actual data.