Reimplemented ReverifyStore to avoid *schema.Table
Using *schema.Table inside ReverifyEntry is an issue because each
DMLEvent stores its own copy of the schema.Table and thus no comparison
between two ReverifyEntry will ever equal true (as is done via the map
key for ReverifyStore). This means reverify entries are NOT
deduplicated in the ReverifyStore.

To fix this, a new struct that stores only the table schema name and the
table name is introduced. We could not use schema.Table as is since it
contains complicated fields and thus is not comparable. The new struct,
TableIdentifier, is used as the map key for ReverifyStore and used for
deduplication. The actual verification processes till needs the
schema.Table object in order to find the pk columns, so the iterative
verifier tries to get the table from the TableSchemaCache, a component
that we pass in as a paramter when initializing the IterativeVerifier.

Further changed the system such that when we build a list of pks as a
batch to verify, we delete the old map based store that allow us to
deduplicate to reduce memory usage. We delete the map partially as we
build the batches so the memory usage of the verifier should never 2x
during this process.

Also added code to emit a logline every time we add 10000 entries into
the ReverifyStore, as a debug log.