WaitUntilReplicaIsCaughtUpToMaster
If we are copying from a slave source database, we must ensure that the
slave is up to date after making the source read only before flushing
and stopping the binlog streamer. Without doing so, a race condition
exists to cause data corruption. As an example:

1. User sets source master writer to READONLY. Suppose master binlog position
   is now 10 and slave binlog position is 8.
2. Ghostferry request for binlog streamer to stop. The binlog streamer
   sets the target stop position 8.
3. Ghostferry misses binlog entry 9 and 10, causing data corruption.

This commit fixes the problem by making sure the replica is caught up to
the master by introducing a optional loop checking this before we flush
and stop the binlog streamer. This is accomplished by:

1. User sets source master writer to READONLY.
2. Ghostferry connects to the master writer and reads the binlog
   position via SHOW MASTER STATUS. This position is saved.
3. Ghostferry connects to the slave and reads the current master binlog
   position via a custom query. As an example, you could SELECT the master
   node from the pt-heartbeat table and thus discover its binlog position
   that has been replicated to the slave.
4. Ghostferry continues to execute this custom query until the position
   is greater than the target position saved in step 2.

We use a heartbeat table because `SHOW SLAVE STATUS` does not give the
correct result if the slave is not replicating directly from the master,
as in a case with a complex replication topology. However, it is
possible to write some code to support the case of `SHOW SLAVE STATUS`
as the interface is general.