Merge ShardingErrorHandler with PanicErrorHandler
Progress on #47.

* HTTPCallback moved from sharding to ghostferry
* HTTPCallback is optional in PanicErrorHandler
* HTTPCallback will only be called once in PanicErrorHandler (was called multiple times if multiple errors occurred simultaneously)

Regarding test coverage changes, ShardingErrorHandler embeds a test ErrorHandler to avoid a panic in tests.
PanicErrorHandler does not allow embedding of additional error handlers, therefore
if invoked in a test, a panic would occur.

We *want* a panic to occur, as that is the intended behavior of PanicErrorHandler,
so we allow the panic in the test, catch it with a defer, and add assertions.

For all other callback tests, we use the test error handler, as before.

We merged the assertions about the HTTP part and the panic part of the callback into a single
test to avoid catching panics in more than one test.