Cancel BinlogStreamer WithTimeout context
We previously discarded the CancelFunc. According to the Go context
documentation[1] (emphasis mine):

> The WithCancel, WithDeadline, and WithTimeout functions take a Context
> (the parent) and return a derived Context (the child) and a CancelFunc.
> Calling the CancelFunc cancels the child and its children, removes the
> parent's reference to the child, and stops any associated timers.
> **Failing to call the CancelFunc leaks the child and its children until
> the parent is canceled or the timer fires.**

This is to say, there will be a goroutine running until cancel is called
or the deadline/timeout has exceeded (see context.go func
propagateCancel). Since this context is a timer based one, the goroutine
will come to a halt eventually. However, this is still not best practise
and thus we can address this here.

[1]: https://golang.org/pkg/context/