context.Context to everything except Verifiers
context.TODO() is used for the ShardingFerry and some testing setup.
These will be addressed in a future commit.

The process I took to make this change is to read the source code from
CopydbFerry.Run down, inspecting each method to see if it needs a
context for cancellation. Currently, we are not passing the context into
the SQL driver and the metrics emitter. This could be addressed in the
future but is not done so for now.