Pass the current BinlogPosition to DMLEvent
Since the BinlogWriter is a separate goroutine to the BinlogStreamer,
the position that it wrote to is important if we want to interrupt and
resume Ghostferry (interrupt saves BinlogWriter position, and resume
starts the BinlogStreamer from the BinlogWriter position).

To allow for this, we need to pass the binlog position from the binlog
streamer, via the DMLEvent, to the binlog writer. This patch adds
BinlogPosition() to DMLEvent to make this possible.