Refactored Ferry to create components
Made BinlogStreamer independent of Ferry.Config

Previously BinlogStreamer was dependant on Ferry.Config. It even went so
far to change Ferry.Config for the automatic MyServerId generation. This
will become problematic as we create new BinlogStreamer instances to
monitor the target database to detect schema changes.

We also exposed methods to create and initialize DataIterator,
BinlogStreamer, and all the major components from Ferry, so another
object that operates at the same level as the Ferry can be created (for
example: the IterativeVerifier, the DeltaCopier, the Reconciler).

The StateTracker was also made optional from the previous commit, as
this does make things kind of easier to work with.