DataIterator.Run(tables)
Instead of passing the Table as a struct member, this makes it pass by
argument in the Run function. It makes more sense as it's not really a
configuration on how to Run the DataIterator, but what to run the
DataIterator on.

This also makes the DataIterator slightly easier to use independently of
the Ferry.