Added some helpers to handle the database.
The DataWriterHelper is a straight port from data_writer.go, which is a
multithreaded data writer that writes random data to the database.

The DbHelper helps connects to the database, seed it, and verify that
the source matches the target so that tests don't have to implement it
over and over again.