Move IterativeVerifier's control into Ferry
This commit moves the initialization and management of IterativeVerifier
from the application level to the core library. This is required to
interrupt and resume the IterativeVerifier as otherwise the application
layer will have to assume the responsibility of resuming the
IterativeVerifier correctly, which is a non-trivial task for the
application level.

This also makes the management of the verifier component cleaner as it
is now an critical part of Ghostferry.

Implication for configurations
------------------------------

Since the verifiers will be initialized by the Ferry using
configurations specified in Config. The Config interface is almost the
same as the previous config, so a small change for sharding is needed:

Before for sharding:

```
{
 "IgnoredVerificationTables": [],
 "IgnoredVerificationColumns": [],
 "MaxExpectedVerifierDowntime": "1m",
 "VerifierIterationConcurrency": 4
}
```

After for sharding:

```
{
  "IterativeVerifierConfig": {
    "IgnoredTables": [],
    "IgnoredColumns": [],
    "MaxExpectedDowntime": "1m"
    "Concurrency": 4
  }
}
```

copydb doesn't require changes because previously it didn't allow for
verifier configurations.

Future improvements
-------------------

One issue that exists is the table schema cache must be given to the
verifiers after Ferry.Start. This causes some ugly code
(Verifier.SetApplicableTableSchemaCache) for now. Once we refactor how
the TableSchemaCache is handled, this method should go away.

Another issue is that the VerifyBeforeCutover is no longer controllable
by the application level. The application level doesn't even really get
an update about this development. This can be addressed in the future.