Fixed JSON column with binlog streamer
JSON column was previously being recognized as a []byte array. This
causes the query generated to be something like `_binary'{"data": 1}'`.
However, JSON data are not binary, thus we get errors such as:

Error 3144: Cannot create a JSON value from a string with CHARACTER SET
binary error.

Fixing this requires us to pass the column types into the methods
actually escaping the values, as we cannot infer that something is a
JSON type from the Go type alone.

To do this the most generic way, I just pass down the table columns all
the way (in case we need something else from it in the future).

Fixes #86.