Use a custom TableSchema struct
Instead of relying on a third party struct for such a crucial piece of
information, we use an extended version of this struct so we can store
custom information on it. The main purpose is to store which columns are
compressed to help with implementing the InlineVerifier.

DMLEvents and RowBatch also no longer hold their own copy of the table
schema structure. They originally held their own copy as we thought
about dealing with schema changes dynamically. Since our present plan is
to pause when a schema change occurs and resume later, we don't need
this anymore. This simplifies the code as all schema structures comes
from the TableSchemaCache.

Lastly, RowBatch.AsSQLQuery and DMLEvent.AsSQLString have been changed
such that they take the schema and table name as string arguments
instead of the *schema.Table struct.