Renamed Status to StatusDeprecated
We want to send Status via a HTTPCallback to some reporting server.
However, the current Status object has too much clutter and is too tied
to the ControlServer.

I don't want to break the ControlServer on master. So I'm moving this to
file to StatusDeprecated and creating a new Status object. Eventually, I
can merge these two implementations back together by creating a struct
specific for the ControlServer and a more generic struct that can be
used everywhere else.