Do not send empty binlog event list to listeners
If all binlog events are filtered out by ApplicableEvents within the
BinlogStreamer, an empty event list will be sent to the event listeners.
This may be unexpected to the event listeners. To save the trouble, if
the BinlogStreamer detects an empty binlog event list, it'll just not
invoke the event listeners.