Rename sharding.Config.SourceDB => SourceSchema
`SourceDB` and `TargetDB` on `sharding.Config` refers to the schema name
of a database. `SourceDB` is also used on `ghostferry.Ferry` to point to
a `*sql.DB`. Since the `Ferry` struct also embeds a `Config` struct (a q
uestionable choice to begin with), this naming caused me significant
confusion.

I'm not sure if this adds to more confusion, but I want to quickly raise
this issue as this is not the first time I got confused by variable
names using words such as: "schema", "db"/"database", "table", and "table
schema".