Additional info logs in DataIterator
If debug level logs are turned off (as in ghostferry-sharding), there
are no visibility of the DataIterator's progress in the logs. This PR
adds at most 50 lines of logs while the DataIterator runs.

The log lines are emitted after a table is queued. Since the queue is an
unbuffered channel and consumed by the tableIterators, logging here will
reliably report the approximate number of tables copied by Ghostferry.