Fix variable names
Implementation of https://github.com/Shopify/ghostferry/pull/138 was a
quick hack to attain the required functionality. This PR serves the sole
purpose of ensuring naming now reflects the usage of `Pagination Column`
as the pivot for iteration during a `Ghostferry` migration.

First my primary focus was to get rid off the deceptive naming off:
`GetPKColumn` and to remove the unneeded index:

    * Change naming GetPKColumn -> GetPaginationColumn"
    * Remove unneeded index

Next the following changes have been made incrementally:

    * PrimaryKey -> PaginationKey
    * quotedPK -> quotedPaginationKey
    * PK -> PaginationKey
    * targetPK -> targetPaginationKey
    * TargetPK -> TargetPaginationKey
    * LastSuccessfulPK -> LastSuccessfulPaginationKey
    * lastSuccessfulPK -> lastSuccessfulPaginationKey
    * completedPK -> completedPaginationKey
    * PKs -> PaginationKeys
    * PKPositionLog -> PaginationKeyPositionLog
    * deltaPK -> deltaPaginationKey
    * modify various PK -> PaginationKey
    * Pk -> PaginationKey
    * pkColumn -> paginationKeyColumn
    * pkTable -> primaryKeyTable
    * bulk update pk -> paginationKey