Fix binary-column data corruption (#157)
THIS PULL REQUEST IS NOT INTENDED TO BE MERGED, JUST FOR INTERNAL
DISCUSSION. WE WILL NEED TO MODIFY THE UPSTREAM VERSION OF THE
VENDOR'ED MODULE INSTEAD OF MODIFYING THE IMPORTED VERSION!

This commit addresses a data corruption bug in the binlog streaming
phase, where the binlog writer incorrectly propagates values in
fixed-length BINARY columns that have trailing 0s in the original value.
These trailing 0s are removed from the binlog by the SQL master and
therefore do not show up in the WHERE clause for update/delete
statements executed by the binlog writer.

Change-Id: I7cd01c953ef2027e4f4fa71a62c5d7b30caa83c5