Unit-test for binary-column data corruption (#157)
There seems to be a bug in the propagation of data from the binlog
streamer to the binlog writer when using fixed-length BINARY columns.

This commit adds an integration test that currently works only under
very specific conditions, and prepares the code for the fix to the data
propagation logic.

Change-Id: Icef7ae558a38af4088280bb633f331711b835502