Allow state tracking via fixed file #164
The previous handling of state across runs of ghostferry was only
allowing the use of stdout. In cases where the tools fails to start,
this mean that the output in stdout was not a valid state, and the
caller has to take special precaution to not override the previous
state.

This commit extends the state handling to allow passing a state file in
the configuration. This state is loaded (if not overridden with the
state file provided as command-line argument), and the state is written
to this file only when the tool shuts down in a way that the output will
be a valid state.

Change-Id: If39728a7c3fc3e0182c2f972f771fb698371d0ce