Introduce "Lock Strategy" option
With this commit, we allow using an in-application lock in ghostferry,
instead of using the source DB as lock. The lock is required to avoid
race conditions between the data iteration/copy and the binlog writer.

The default behavior is preserved; a new option "LockStrategy" allows
moving the lock from the source DB into ghostferry, or disabling the
lock altogether.

This fixes #169

Change-Id: I20f1d2a189078a3877f831c7a98e8ca956620cc7