Reproduces uint panic in InlineVerifier
See https://github.com/Shopify/ghostferry/issues/165.

The issue appears to be when the PK column is `BIGINT UNSIGNED`. For
some reason I was under the impression that that column type is what we
used internally. It turns out we are mostly using ghostferry with
dataset where the PK is always `BIGINT` (signed).

If you look through the test, we also don't really use `UNSIGNED`
anywhere except in `TestCopyDataWithLargePaginationKeyValues`. However,
this test does not attempt to insert into the source while copy is
happening. This therefore does not trigger the bug, as it is only
revealed in DML events.