Fixed uint64 panic in InlineVerifier
Fixes https://github.com/Shopify/ghostferry/issues/165.

So the reason we were all super confused is because I falsely assumed
that we run Ghostferry against `bigint unsigned` most of the time while
we actually didn't. Also adding to the confusion is because `GetUint64`
is processing value received from the DataIterator (via
go-sql-driver/mysql) and BinlogStreamer (via siddontang/mysql). This
fact was not immediately recognized and we falsely assumed that the code
as is should be correct. Since the original code was only designed to
be used for the DataIterator, the unexpected type from siddontang/mysql
causes a panic.

Co-authored-by: Clemens Kolbitsch <kolbitsch@lastline.com>
Co-authored-by: Daniel Oliveira <daniel.oliveira079@gmail.com>