Moved VerifyBeforeCutover after wait for cutover
As it stands, the inline verifier periodic reverifier will be stopped
the moment data copy is complete. It could be hours before we wake up
and perform the real cutover. During these hours, the backlog of binlogs
to be verified can grow to enormous amount, which can cause
unbounded/very large verification time during cutover, which can result
in extended downtime.