TLA+ interrupt-resume with DataIterator
Instead of interrupting and then resuming in TLA as that would be messy
and difficult to check, we generate starting states for the SourceTable
and TargetTable such that it looks like it was previously interrupted.
This is a multiple-starting state approach that significantly increases
the search space of the TLC run.

We also generate a valid "LastSuccessfulPK" that will be valid given the
table states such that some of the initial states will cause the
Ghostferry algorithm to re-copy previously copied rows. There's no
mathematical proof that as long as we can re-copy one row correctly, we
can re-copy n rows correctly, but intuition says this is indeed the
case.