TLA+ interrupt-resume with BinlogStreamer
Taking a similar approach with the TLA+ for the DataIterator, the idea
here is we generate a possible SourceBinlog sequence at the initial
starting state.

This time, we generate a SourceBinlog that's either an empty sequence
(simulating a fresh Ghostferry run) or contains a single binlog entry
that has already been previously copied over to the TargetTable. The
TLA+ specification is constructed such that when Ghostferry starts, it
will attempt to copy that binlog over to the target again. Effectively,
this simulates when Ghostferry is resumed from a binlog coordinate
that is "outdated": i.e. if a binlog is already copied but the last
successful binlog position is not updated, it is safe to resume from an
outdated position.

Once again, this is only a single binlog entry and we do not have a
mathematical proof that if it works with a single binlog entry, it
works with n binlog entries. However, strong intuitions suggests that
this is the case so this proof is omitted for now.