REPLACE INTO and INSERT IGNORE is the same thing for DataIterator
I noticed that the original implementation of Ghostferry's TLA+, the
DataIterator is implement in a way such that the data is copied with
equivalent of a REPLACE INTO instead of INSERT IGNORE. The TLC results
suggests both of these are fine, with or without interrupt/resume. At
first glance this feels unintuitive, shouldn't the REPLACE INTO variant
be just wrong and cause data corruption? After much thought, I'm fairly
certain that this is expected, and it is better illustrated with the
following state table, where each subsequent row is one logical timestep
forward

**Case 1**:

+---------------+------------------+------------------+
| Operation     | Source Row State | Target Row State |
+---------------+------------------+------------------+
| Start         | V1               | nil              |
| Data Copy     | V1               | V1               |
| Source Update | V2               | V1               |
| Binlog Apply  | V2               | V2               |
+---------------+------------------+------------------+

This case is when the data copy occurs before the source update. Note
that since we use SELECT FOR UPDATE, the source row cannot be updated
until the data copy (via either REPLACE INTO or INSERT IGNORE) is done.
Which means it's impossible that during Data Copy we read V1 and then
somehow Binlog Apply fires before we perform the REPLACE INTO.

**Case 2**:

+---------------+------------------+------------------+
| Operation     | Source Row State | Target Row State |
+---------------+------------------+------------------+
| Start         | V1               | nil              |
| Source Update | V2               | nil              |
| Binlog Apply  | V2               | V2               |
| Data Copy     | V2               | V2               |
+---------------+------------------+------------------+

Replacing Data Copy with either an INSERT IGNORE or a REPLACE INTO
accomplishes the same. At the end of the 3rd step, the target row will
always be V2.

**Case 3**:

+---------------+------------------+------------------+
| Operation     | Source Row State | Target Row State |
+---------------+------------------+------------------+
| Start         | V1               | nil              |
| Source Update | V2               | nil              |
| Data Copy     | V2               | V2               |
| Binlog Apply  | V2               | V2               |
+---------------+------------------+------------------+

This case is simply putting BinlogApply after the Data Copy, which is
fine because Binlog Writer is using either INSERT IGNORE or a precise
WHERE match to update rows, which results in a no-op.

**Case 4**:

In this case, we expand Binlog Apply into two operations that is
interlaced with Data Copy. So Case 2 and Case 3 are really just with
Binlog Read + Binlog Write grouped together before and after the Data
Copy.

+---------------+------------------+------------------+
| Operation     | Source Row State | Target Row State |
+---------------+------------------+------------------+
| Start         | V1               | nil              |
| Source Update | V2               | nil              |
| Binlog Read   | V2               | nil              |
| Data Copy     | V2               | V2               |
| Binlog Write  | V2               | V2               |
+---------------+------------------+------------------+

Again, this is fine because the Binlog Write operation will only write
if the target row state is V1. Since it's already V2, it is a no-op.

**Other Cases**?

There should be no other cases. The constraints is that we have to have
Source Update before Binlog Apply can occur, whereas Data Copy can occur
at any point in time.