False positive InlineVerifier errors
This commit fixes a false positive problem with the InlineVerifier,
which is documented in #149. That bug report documented a relatively
complex and uncommon race condition that could occur and result in a
false positive inline verifier error. However, with the presence of
interrupt resume with periodic snapshots in the past, this problem
becomes a lot more prevalent. An example case is given in the test, and
is also given below for reference:

| Step | Action             | Source | Target | Notes |
|------|--------------------|--------|--------|-------|
| 0    | Initial State      | v1     | nil    |       |
| 1    | State Dump         | v1     | nil    |       |
| 2    | Copy row           | v1     | v1     |       |
| 3    | Interrupt          | v1     | v1     |       |
| 4    | App UPDATE         | v2     | v1     |       |
| 5    | Resume from step 1 | v2     | v1     |       |
| 6    | Copy row           | v2     | v1     | NO-OP due to insert ignore. |
| 7    | CHECKSUM           | v2     | v1     | InlineVerifier error here.  |
| 8    | Process halted     | v2     | v1     |       |

This means that if a move fail like this, it might never resume, as the
binlog streamer could be quite lagged and therefore the error will
repeat itself on every resume. Suppose if we do not error immediately on
step 7 and allow the process to continue, the binlog streamer will
eventually catch up and correct that row before cutover. This checksum
table error could be queued in the background reverify store and
periodically reverified until either the source and target matches, or
until cutover, when it would fail if the discrepency is not resolved.

The following example is a run after this commit:

| Step | Action             | Source | Target | Notes |
|------|--------------------|--------|--------|-------|
| 0    | Initial State      | v1     | nil    |       |
| 1    | State Dump         | v1     | nil    |       |
| 2    | Copy row           | v1     | v1     |       |
| 3    | Interrupt          | v1     | v1     |       |
| 4    | App UPDATE         | v2     | v1     |       |
| 5    | Resume from step 1 | v2     | v1     |       |
| 6    | Copy row           | v2     | v1     | NO-OP due to insert ignore. |
| 7    | CHECKSUM           | v2     | v1     | Row added to reverify store.|
| 8    | Binlog apply       | v2     | v2     |       |
| 9    | Reverify           | v2     | v2     | Verification successful, row is dropped from reverify store |

If step 9 is not successful here because step 8 failed to apply
correctly, resulting in a data corruption, the row will not be dropped
from the reverify store. During cutover, a final verification will be
performed, and the Ghostferry process will error due to a verification
error.