Added binlog verify store entry count
The currentRowCount emitted before can count the same row multiple
times, which does not allow for a great estimate of how big the state
dump could be. With this patch, we emit both the row count (which is
maybe useful to estimate how long it would take to reverify) and the
entries count (which is useful to estimate how big the statedump is) in
the VerifierMessage that's dumped to the progress.