Fixed memory leak regression in binlog streamer
The binlog streamer is canceling the context with a defer. This defer is
called in a loop, which means the deferred function will not be called
until the loop exits. For Ghostferry, that loop only exit at the very
end, which means we have a memory leak with all the deferred function
calls.