Fixed Incorrect arguments to mysqld_stmt_execute
This error appears to only happen on CI and not locally, despite
everything running with the same version. I have no idea why this is the
case.

Since I removed the `nil` from the prepared statement, it is possible
that this problem is related to the issues linked below, but I'm not sure. I also
don't really know how to debug in github actions yet, so I don't know
how to confirm if this is indeed the case.

- https://github.com/go-sql-driver/mysql/issues/209
- https://github.com/launchbadge/sqlx/issues/125
- https://stackoverflow.com/questions/16726752/sending-null-in-a-prepared-statement-for-mysql-doesnt-seem-to-work?rq=1