Code that makes tests pass
Tldr;

The primary change in this commit has the following aim: Detect when the schema for a table has changed on the source
and refresh the schema cache. A longer explanation follows.

The long version:

Ghostferry caches the schemas of all tables it's asked to track when it starts. It compares every binlog event that
comes in with the cached schema (using column length) to make sure that the schema has not changed and fails if it has.

We need to make Ghostferry aware of these schema changes and let it gracefully update its cache if we want it to deal
with LHMs. To do this, we follow the following strategy:
1. We maintain an additional "table ID" cache which maps each fully qualified table name to its table ID in the current
binlog session.
2. For each RowEvent, we check the associated TableMapEvent and verify that the schema in the event matches that cached
for that table name in our cache. We do this by comparing the table IDs. Ideally we'd compare the list of columns and
their types, but column names are only available in MySQL 8.0.1+, which our source DBs won't be running. So this
approach is an approximation, based on the assumption that table IDs change when the schema changes.
3. When we detect that the table ID for a given table has changed, we reload the schema for that table only, refresh the
table schema cache, then refresh the table ID cache with the new table ID.

This allows us to gracefully handle events that follow the new schema without panicing. This approach does not work for
DataIterators, though, because each data iterator is a goroutine that starts with a static list of table schemas, and
cannot gracefully handle schema changes.