Make data iterator work with column names + formatting etc
Data iterators use some of the same verification logic as binlog streamer. However, while binlog streamer has access to
the latest table schema after the event being refreshed, data iterators don't, because they are only looking at the
table schema that the goroutine was started with. To approximate "live" events, we infer the table schema from the
number of columns in the rows read instead.